Make Program.cs for the CLI only
