+++ src/SourceBuild/content/eng/tools/BuildComparer/BaselineEntry.cs
@@ -1,4 +1,7 @@
-﻿using System.Text.RegularExpressions;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Text.RegularExpressions;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Xml.Serialization;
